BUG: Hidden files are not being present in the docs/demo directory
Files starting with a dot that are included in "docs/files" are not
being added to the "demo" folder.

JSDoc configuration suffers from this bug.
CopyWebpackPlugin seems to be suffering this too.

SOLUTION:

Remove files from "docs/files" directory that start with a dot ".", like ".gitignore".

Exclude those files in the "update-gh-page" script and commit those
files directly to the gh-pages branch.