BUG: Fix "webpack bundles not found in docs"
Add a plugin to copy files.
On production, copy bundles to "docs/demo" directory.

CAUSE:

Webpack bundles weren't being saved in the docs directory for production
mode, so "npm run build" wasn't working properly.

For development mode that wasn't the case, since the default directory
for serving bundles is in that path.