MAINT: Use "jsdoc" as dependency (instead of dev dependency)
Remove tracking branch.

REASON:
Code was being updated manually with a tracking branch, but being up to
date with the master branch is actually not important. By using jsdoc as
a dependency, original code will be in node_modules and changes introduced
by this template will be compared with the original code, revised and
committed to the master branch. By doing so, the template will use the
latest released version and contain only one remote.

MAINT: Move dependencies to devDependencies