STY: Remove ugly aspects of the template, like things enclosed with "<>"
Replace things like 'Array.<string>' with 'string[]' (except in headings).
Replace '*' and '|' with 'Any' and 'or', respectively.
Replace '&hellip;' with '...'.

STY: Remove "Attributes" column to save some space in tables

Add attributes to the "name" field.

STY: Add :hover animation to examples and change color to .code-caption

Use the same color in .code-caption as the source code.

STY: Modify text color in the prettify theme